+++ temba/channels/models.py
@@ -194,6 +194,7 @@ class Channel(TembaModel):
     CONFIG_PASSWORD = 'password'
     CONFIG_KEY = 'key'
     CONFIG_API_ID = 'api_id'
+    CONFIG_API_KEY = 'api_key'
     CONFIG_CONTENT_TYPE = 'content_type'
     CONFIG_VERIFY_SSL = 'verify_ssl'
     CONFIG_USE_NATIONAL = 'use_national'
@@ -1412,6 +1413,7 @@ class ChannelEvent(models.Model):
     TYPE_NEW_CONVERSATION = 'new_conversation'
     TYPE_REFERRAL = 'referral'
     TYPE_FOLLOW = 'follow'
+    TYPE_STOP_CONTACT = 'stop_contact'
 
     EXTRA_REFERRER_ID = 'referrer_id'
 
@@ -1421,6 +1423,7 @@ class ChannelEvent(models.Model):
                    (TYPE_CALL_OUT_MISSED, _("Missed Outgoing Call"), 'call-out-missed'),
                    (TYPE_CALL_IN, _("Incoming Call"), 'call-in'),
                    (TYPE_CALL_IN_MISSED, _("Missed Incoming Call"), 'call-in-missed'),
+                   (TYPE_STOP_CONTACT, _("Stop Contact"), 'stop-contact'),
                    (TYPE_NEW_CONVERSATION, _("New Conversation"), 'new-conversation'),
                    (TYPE_REFERRAL, _("Referral"), 'referral'),
                    (TYPE_FOLLOW, _("Follow"), 'follow'))
@@ -1478,6 +1481,7 @@ class ChannelEvent(models.Model):
         Handles takes care of any processing of this channel event that needs to take place, such as
         trigger any flows based on new conversations or referrals.
         """
+        from temba.contacts.models import Contact
         from temba.triggers.models import Trigger
         handled = False
 
@@ -1490,6 +1494,13 @@ class ChannelEvent(models.Model):
 
         elif self.event_type == ChannelEvent.TYPE_FOLLOW:
             handled = Trigger.catch_triggers(self, Trigger.TYPE_FOLLOW, self.channel)
+
+        elif self.event_type == ChannelEvent.TYPE_STOP_CONTACT:
+            user = get_anonymous_user()
+            contact = Contact.get_or_create(self.org, user, name=None, urns=[self.contact_urn.urn],
+                                            channel=self.channel)
+            contact.stop(user)
+            handled = True
 
         return handled